Move datanode test/wait packages to server
