+++ data-node/src/test/java/org/graylog/datanode/integration/DatanodeSearchableSnapshotsIT.java
@@ -19,8 +19,8 @@ package org.graylog.datanode.integration;
 import com.github.rholder.retry.RetryException;
 import io.restassured.http.ContentType;
 import io.restassured.response.ValidatableResponse;
-import org.graylog.datanode.restoperations.DatanodeOpensearchWait;
-import org.graylog.datanode.restoperations.RestOperationParameters;
+import org.graylog.testing.restoperations.DatanodeOpensearchWait;
+import org.graylog.testing.restoperations.RestOperationParameters;
 import org.graylog.datanode.testinfra.DatanodeContainerizedBackend;
 import org.graylog.testing.completebackend.S3MinioContainer;
 import org.graylog.testing.containermatrix.MongodbServer;
