+++ graylog2-server/src/test/java/org/graylog/testing/restoperations/DatanodeRestApiWait.java
@@ -14,7 +14,7 @@
  * along with this program. If not, see
  * <http://www.mongodb.com/licensing/server-side-public-license>.
  */
-package org.graylog.datanode.restoperations;
+package org.graylog.testing.restoperations;
 
 import com.github.rholder.retry.RetryException;
 import io.restassured.response.ValidatableResponse;
